backend/requirements.txt
protobuf~=5.29.4
pypdf2~=3.0.1
python-docx~=1.1.2
python-dotenv~=1.1.0
requests~=2.32.3
pydantic~=2.11.2
google-generativeai~=0.8.4
fastapi
sqlalchemy.orm
psycopg2
pydantic[email]
passlib
python-jose[cryptography]
python-multipart
uvicorn
aiofiles
sentence-transformers~=4.0.2
ipykernel~=6.29.5
notebook~=7.4.0
pandas
numpy
pymongo
scikit-learn
torch
spacy